arm-v7m: execute syscalls in supervisor thread context
Syscalls shouldn't be executed in interrupt context, because it prevent
the syscall from sleeping.

This commit change the svcall interrupt handler in order to return as soon
as possible. The interrupt handler modify its stack frame in order to return
to a syscall handler that run in supervisor thread mode and in the context of
the task that called the syscall. This function will do the actual syscall
task and then will return to the function that made the syscall.

Signed-off-by: Fabien Parent <parent.f@gmail.com>